requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-ollama
faiss-cpu
pypdf
python-dotenv
sentence-transformers